back/users/requirements.txt
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
asyncpg==0.30.0
click==8.2.1
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.116.2
greenlet==3.2.4
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tomli==2.2.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
